mpi/init: move finalize callback util to init_util.c
Move the utility code (just the callback utility for now) to separate
file -- init_util.c.

Rename MPIR_Call_finalize_callbacks to MPII_Call_finalize_callbacks
since it is only used by MPIR_Finalize_impl internally and is
consistenty with other routines in mpi_init.h.
